perf(telegram): isolate migration persistence metadata
